Create Alternating Characters
You are given a string containing characters A and B only, your task is to change it into a string such that every two consecutive characters are different. To do this, you are allowed to delete one or more characters in the string.

Your task is to find the minimum number of required deletions.

For example, string ABAA should be changed to ABA by deleting one character A.

Input Format

The first line contains an integer <t>, i.e. the number of test cases. 
The next <t> lines contain a string <s>.

Output Format

For each test case, print the minimum number of deletions required in a new line.