Create Left Rotation
A left rotation operation on an array of size N shifts each of the array's elements 1 unit to the left.
For example, if 2 left rotations are performed on array of size [1,2,3,4,5] then the array would become 
[3,4,5,1,2].

Given an array of N integers and a number, D, perform D left rotations on the array.
Then print the updated array as a single line of space-separated integers.

Input Format

The first line contains two space-separated integers denoting the respective values of N (the number of integers)
and D (the number of left rotations you must perform). The second line contains N space-separated integers 
describing the respective elements of the array's initial state.

Output Format

Print all <N> integers in <A> in reverse order as a single line of space-separated integers.